Update archive_folders_to_standby.md
When using instructions --remove-files option is not correctly placed. `-f` requires a file name after so if written as indicated it will create the --remove-files file instead.
Also, adjusted remove the date from the original folder name as it should not have it in a real case. Date goes with the tar.gz file, not the folder.